+++ lib/src/xform.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/material.dart';
 import 'xfocusnode.dart';
 import 'xformcontainer.dart';
@@ -5,21 +7,28 @@ import 'xformcontainer.dart';
 class XForm extends StatefulWidget {
   final Function nextPage;
   final Function prevPage;
-  final Function submit;
+  final Function(Map<String, dynamic>) submit;
+  final Function(Map<String, dynamic>, bool) onSave;
   final bool autoValidate;
+  final bool autoFocus;
+  final bool autoPageSwitch;
   final bool showButtons;
   final List<Widget> children;
   const XForm(
       {Key key,
       this.children,
       @required this.submit,
+      this.onSave,
       this.nextPage,
       this.prevPage,
       this.autoValidate = false,
+      this.autoPageSwitch = true,
+      this.autoFocus = true,
       this.showButtons = true})
       : super(key: key);
   static XFormState of(BuildContext context) {
-    final _XFormScope scope = context.inheritFromWidgetOfExactType(_XFormScope);
+    final _XFormScope scope =
+        context.dependOnInheritedWidgetOfExactType<_XFormScope>();
     return scope?._formState;
   }
 
@@ -33,37 +42,53 @@ class XFormState extends State<XForm> {
   Map<String, dynamic> formValues = {};
   bool autoValidate;
   List<FocusNode> _focusNodes = [];
-  int page = 1;
+  FocusNode currentFocus;
+  bool switching = false;
+  int page;
   int pages;
 
   @override
   void initState() {
     super.initState();
-    pages = widget.children.length;
+    if (widget.children == null) {
+      page = 0;
+      pages = 0;
+    } else {
+      pages = widget.children.length;
+      page = pages > 0 ? 1 : 0;
+    }
     autoValidate = widget.autoValidate;
   }
 
   prevPage() {
     _formKey.currentState.save();
     if (page != 1) {
-      setState(() {
-        page = page - 1;
-      });
+      if (_formKey.currentState.validate()) {
+        setState(() {
+          autoValidate = widget.autoValidate;
+          page = page - 1;
+        });
+      } else {
+        setState(() {
+          autoValidate = true;
+        });
+      }
       _controller.animateToPage(page - 1,
-          curve: Curves.ease, duration: Duration(milliseconds: 200));
+          curve: Curves.ease, duration: Duration(milliseconds: 250));
     }
   }
 
-  nextPage() {
+  nextPage({int i}) {
     _formKey.currentState.save();
     if (page != pages) {
       if (_formKey.currentState.validate()) {
         setState(() {
-          autoValidate = false;
+          autoValidate = widget.autoValidate;
           page = page + 1;
         });
         _controller.animateToPage(page - 1,
-            curve: Curves.ease, duration: Duration(milliseconds: 200));
+            curve: Curves.ease, duration: Duration(milliseconds: 250));
+        if (i != null) FocusScope.of(context).requestFocus(_focusNodes[i + 1]);
       } else {
         setState(() {
           autoValidate = true;
@@ -72,14 +97,23 @@ class XFormState extends State<XForm> {
     } else {
       submit();
     }
+    cancelSwitch();
+  }
+
+  cancelSwitch() async {
+    await Future.delayed(Duration(milliseconds: 500));
+    switching = false;
   }
 
   register(name) {
     XFocusNode xFocus = XFocusNode();
-    if (_focusNodes.length == 0) {
-      xFocus.autoFocus = true;
-    }
     xFocus.focus = FocusNode();
+    if (switching || widget.autoFocus && _focusNodes.length == 0) {
+      xFocus.autoFocus = true;
+      currentFocus = xFocus.focus;
+      FocusScope.of(context).requestFocus(currentFocus);
+      switching = false;
+    }
     _focusNodes.add(xFocus.focus);
     if (formValues.containsKey(name)) {
       xFocus.defaultValue = formValues[name];
@@ -93,15 +127,25 @@ class XFormState extends State<XForm> {
     });
     if (_formKey.currentState.validate()) {
       _formKey.currentState.save();
+      _focusNodes.last.unfocus();
       widget.submit(formValues);
     }
+  }
+
+  bool save() {
+    _formKey.currentState.save();
+    return _formKey.currentState.validate();
   }
 
   void onSave(field, value) {
     formValues.addEntries([new MapEntry(field, value)]);
   }
 
-  void next(focusNode) {
+  void next(FocusNode focusNode) {
+    currentFocus = focusNode;
+    focusNode.unfocus();
+    FocusScope.of(context).unfocus();
+    switching = true;
     int i = _focusNodes.lastIndexOf(focusNode);
     if (i != _focusNodes.length - 1) {
       FocusScope.of(context).requestFocus(_focusNodes[i + 1]);
@@ -109,24 +153,26 @@ class XFormState extends State<XForm> {
       if (page == pages) {
         submit();
       } else {
-        nextPage();
+        if (widget.autoPageSwitch) nextPage(i: i);
+        if (widget.onSave != null)
+          widget.onSave(formValues, _formKey.currentState.validate());
       }
     }
   }
 
   _buildButtons(int pages) {
     var backButton = RaisedButton(
-      child: Text("Back"),
+      child: Text("Prev"),
       onPressed: page != 1 ? () => prevPage() : null,
     );
     var nextButton = RaisedButton(
       child: Text("Next"),
-      color: Theme.of(context).primaryColorLight,
+      // color: Theme.of(context).primaryColorLight,
       onPressed: () => nextPage(),
     );
     var submitButton = RaisedButton(
       child: Text("Submit"),
-      color: Theme.of(context).primaryColorLight,
+      // color: Theme.of(context).primaryColorLight,
       onPressed: () => submit(),
     );
     List<Widget> buttons = [];
@@ -141,22 +187,28 @@ class XFormState extends State<XForm> {
         buttons.add(nextButton);
       }
     }
-    return Row(
-        mainAxisAlignment: MainAxisAlignment.spaceEvenly, children: buttons);
+    if (pages > 0)
+      return Row(
+          mainAxisAlignment: MainAxisAlignment.spaceEvenly, children: buttons);
   }
 
   @override
   Widget build(BuildContext context) {
+    widget.children.map((c) {
+      if (c is ListView) {}
+    });
     return XFormContainer(
       register: (name) => register(name),
       onSave: (field, value) => onSave(field, value),
       next: (focusNode) => next(focusNode),
+      save: () => save(),
       child: Form(
           key: _formKey,
           autovalidate: autoValidate,
           child: Column(
             children: <Widget>[
               Flexible(
+                fit: FlexFit.loose,
                 child: PageView(
                     physics: ScrollPhysics(),
                     controller: _controller,
@@ -170,6 +222,12 @@ class XFormState extends State<XForm> {
     );
   }
 }
+// widget.children.length == 1
+//                           ? Column(children: widget.children)
+//                           : PageView(
+//                               physics: ScrollPhysics(),
+//                               controller: _controller,
+//                               children: widget.children),
 
 class _XFormScope extends InheritedWidget {
   const _XFormScope({Key key, Widget child, XFormState formState})