+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [0.3.0] - 15 Jul 2020
+* Added onSave callback
+* Added field icon
+* Added text capitalization toggle
+* Added autocorrect/suggestions toggle
+* Added auto pageswitch/autofocus toggle
+* Fixed phone number validation
+* Added multiline for textarea
+* Fixed focus issue on page change
+* Changed date field to activate on tap
+* Added disable field for textfield
+* Bug fixes
+
+## [0.2.2] - 12 Jun 2019
+* Changed the select fields selected option to string [**BREAKING]
+* Fix: Select field default value was not working
+* Fix: Text field returns value on blur/submit
+
 ## [0.2.0] - 9 Apr 2019
 * Date field returns date type
 * Numeric field returns numeric value
@@ -5,12 +23,12 @@
 * Validation on page switch
 
 ## [0.1.2] - 29 Jan 2019
-* Fixed Form display on long forms to not overlap buttons
+* Fix: Form display on long forms to not overlap buttons
 * Added support for checkbox, checkboxgroup and radiogroup
 * Changed default values from string to dynamic
 
 ## [0.1.1] - 17 Jan 2019
-* Fixed bug in date field default values
+* Fix: date field default values
 
 ## [0.1.0] - 16 Jan 2019
 * Added styling options to fields